Update SK's HDR enabled detection code to account for WCG SDR displays, as Windows 11 24H2 finally makes public the API to distinguish WCG from HDR.

+++ src/control_panel.cpp
@@ -4908,7 +4908,7 @@ SK_ImGui_ControlPanel (void)
 
       if (SK_IsGameWindowActive ())
       {
-        if (effective_power_mode != EffectivePowerModeNone)
+        if ((int)effective_power_mode != EffectivePowerModeNone)
         {
           ImGui::SameLine    ();
           ImGui::SeparatorEx (ImGuiSeparatorFlags_Vertical);
